Organize broker wire codec by protocol layer
